Minor bug fixes for WiFire build

+++ core/src/common/lwm2m_tree_builder.c
@@ -59,7 +59,8 @@ static AwaResult ReadResourceInstanceFromStoreAndCreateTree(Lwm2mTreeNode ** des
         goto error;
     }
 
-    Lwm2m_Debug("Treebuilder length: %zd\n", valueLength);
+    //Lwm2m_Debug("Treebuilder length: %zd\n", valueLength);
+    Lwm2m_Debug("Treebuilder length: %d\n", (int)valueLength);
 
     if (Lwm2mTreeNode_SetValue(*dest, (const uint8_t*)value, valueLength) != 0)
     {
